Fixed state change of payment_status and payment_balance

+++ database/migrations/2024_07_12_115053_add_payment_balance_to_reports_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('reports', function (Blueprint $table) {
-            $table->float('payment_balance', 8, 2)->nullable();
+            $table->decimal('payment_balance', 10, 2)->nullable();
         });
     }
